Refactoring Outputs to implement a standard interface
