+++ lib/output/output.go
@@ -1,11 +1,8 @@
 package output
 
-import (
-	"encoding/json"
-	"fmt"
-	"sort"
-	"strings"
-)
+type OutputBuilder interface {
+	GenerateOutput(interface{}) (string, error)
+}
 
 type Config struct {
 	Format string
@@ -20,111 +17,21 @@ func (o *Output) Generate(elements interface{}) (string, error) {
 		return "", nil
 	}
 
+	var v OutputBuilder
 	switch o.Config.Format {
 	case "json":
-		return o.JSONOutput(elements)
+		v = &JSONOutput{
+			Config: o.Config,
+		}
 	case "shell":
-		return o.ShellOutput(elements)
-	}
-
-	return "", fmt.Errorf("Unrecognized output format: %s", o.Config.Format)
-}
-
-func (o *Output) JSONOutput(elements interface{}) (string, error) {
-	if _, ok := elements.([]interface{}); ok {
-		if j, err := json.MarshalIndent(elements, " ", " "); err != nil {
-			return "", err
-		} else {
-			return string(j), nil
-		}
-	} else {
-		if _, ok := elements.(map[string]interface{}); ok {
-			if j, err := json.MarshalIndent(elements, " ", " "); err != nil {
-				return "", err
-			} else {
-				return string(j), nil
-			}
-		} else {
-			return fmt.Sprintf("%s", elements), nil
-		}
-	}
-}
-
-func (o *Output) ShellOutput(elements interface{}) (string, error) {
-	var keys []string
-	var output string
-
-	parsed := make(map[string]string)
-	_, _, parsed = parseElementsForShell("elements", elements, parsed)
-
-	for k := range parsed {
-		keys = append(keys, k)
-	}
-
-	sort.Strings(keys)
-
-	for _, k := range keys {
-		output = fmt.Sprintf("%s%s=\"%s\"\n", output, k, parsed[k])
-	}
-
-	return output, nil
-}
-
-func parseElementsForShell(label string, elements interface{}, parsed map[string]string) (string, interface{}, map[string]string) {
-	var e interface{}
-	switch elements.(type) {
-	case map[string]interface{}:
-		for key, subElements := range elements.(map[string]interface{}) {
-			l := fmt.Sprintf("%s_%s", label, sanitizeKeyForShell(key))
-			_, e, parsed = parseElementsForShell(l, subElements, parsed)
-		}
-	case []interface{}:
-		for i, subElements := range elements.([]interface{}) {
-			l := fmt.Sprintf("%s_%d", label, i)
-			_, e, parsed = parseElementsForShell(l, subElements, parsed)
+		v = &ShellOutput{
+			Config: o.Config,
 		}
 	default:
-		parsed[label] = fmt.Sprintf("%v", elements)
-		return label, nil, parsed
+		v = &InvalidOutput{
+			Config: o.Config,
+		}
 	}
 
-	return label, e, parsed
-}
-
-func sanitizeKeyForShell(key string) string {
-	// replace non-compliant shell characters with underscores,
-	// per IEEE Std 1003.1-2001
-	r := strings.NewReplacer(
-		"`", "_",
-		"~", "_",
-		"!", "_",
-		"@", "_",
-		"#", "_",
-		"$", "_",
-		"%", "_",
-		"^", "_",
-		"&", "_",
-		"*", "_",
-		"(", "_",
-		")", "_",
-		"-", "_",
-		"+", "_",
-		"=", "_",
-		"{", "_",
-		"}", "_",
-		"[", "_",
-		"]", "_",
-		"|", "_",
-		":", "_",
-		";", "_",
-		"'", "_",
-		"\"", "_",
-		"/", "_",
-		"?", "_",
-		",", "_",
-		".", "_",
-		"<", "_",
-		">", "_",
-	)
-	return r.Replace(key)
+	return v.GenerateOutput(elements)
 }
